+++ tests/integ/scala/test_dataframe_reader_suite.py
@@ -7,6 +7,7 @@ import logging
 import random
 from decimal import Decimal
 from unittest import mock
+import json
 
 import pytest